+++ src/store/modules/coaches/mutations.js
@@ -1,0 +1,8 @@
+export default {
+  addNewCoach(state, payload) {
+    state.coaches.push(payload)
+  },
+  setCoaches(state, payload) {
+    state.coaches = payload;
+  }
+}
